backend/requirements.txt
Django==2.2.16
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0
drf-extra-fields==3.1.1
flake8==5.0.4
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.19.0
djoser==2.1.0
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.10
